Refactor and split up files, break up strategies and routes
