+++ package.json
@@ -24,29 +24,6 @@
     "@bugsnag/js": "^6.0.0",
     "@bugsnag/plugin-express": "^6.0.0",
     "@sendgrid/mail": "^6.3.1",
-    "@types/cheerio": "^0.22.11",
-    "@types/compression": "0.0.36",
-    "@types/connect-flash": "0.0.34",
-    "@types/connect-mongo": "0.0.41",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/cookie-signature": "^1.0.2",
-    "@types/express": "^4.16.1",
-    "@types/express-session": "^1.15.12",
-    "@types/html-to-text": "^1.4.31",
-    "@types/marked": "^0.6.3",
-    "@types/moment-timezone": "^0.5.12",
-    "@types/mongoose": "^5.3.21",
-    "@types/morgan": "^1.7.35",
-    "@types/node": "^11.11.0",
-    "@types/oauth2orize": "^1.8.5",
-    "@types/passport": "^1.0.0",
-    "@types/passport-facebook": "^2.1.8",
-    "@types/passport-github2": "^1.2.4",
-    "@types/passport-http": "^0.3.7",
-    "@types/passport-http-bearer": "^1.0.33",
-    "@types/passport-local": "^1.0.33",
-    "@types/passport-oauth2-client-password": "^0.1.2",
-    "@types/uuid": "^3.4.4",
     "bugsnag-build-reporter": "^1.0.2",
     "chalk": "^2.4.2",
     "cheerio": "^1.0.0-rc.2",
@@ -81,5 +58,30 @@
     "typescript": "^3.3.0",
     "uuid": "^3.3.2"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@types/cheerio": "^0.22.11",
+    "@types/compression": "0.0.36",
+    "@types/connect-flash": "0.0.34",
+    "@types/connect-mongo": "0.0.41",
+    "@types/cookie-parser": "^1.4.1",
+    "@types/cookie-signature": "^1.0.2",
+    "@types/express": "^4.16.1",
+    "@types/express-session": "^1.15.12",
+    "@types/html-to-text": "^1.4.31",
+    "@types/marked": "^0.6.3",
+    "@types/moment-timezone": "^0.5.12",
+    "@types/mongoose": "^5.3.21",
+    "@types/morgan": "^1.7.35",
+    "@types/node": "^11.11.0",
+    "@types/oauth2orize": "^1.8.5",
+    "@types/passport": "^1.0.0",
+    "@types/passport-facebook": "^2.1.8",
+    "@types/passport-github2": "^1.2.4",
+    "@types/passport-google-oauth20": "^2.0.4",
+    "@types/passport-http": "^0.3.7",
+    "@types/passport-http-bearer": "^1.0.33",
+    "@types/passport-local": "^1.0.33",
+    "@types/passport-oauth2-client-password": "^0.1.2",
+    "@types/uuid": "^3.4.4"
+  }
 }
