+++ src/templates/authorize.hbs
@@ -1,0 +1,61 @@
+{{#> main }}
+    <section class="columns">
+        <div class="column is-half is-offset-one-quarter" id="authorize">
+            <span class="icon is-large">
+                <i class="fas fa-3x fa-user-astronaut"></i>
+            </span>
+
+            <h1 class="title">{{appName}} would like to connect to your account</h1>
+            <p>
+                <span class="icon is-medium"><i class="fas fa-lg fa-satellite"></i></span>
+                Authorizing will redirect to <strong>{{redirect}}</strong>
+            </p>
+            <p>
+                <span class="icon is-medium" style="transform: rotate(270deg)"><i class="fas fa-lg fa-space-shuttle"></i></span>
+                You're logged in as <strong>{{name}}</strong> ({{email}})
+            </p>
+            <br />
+
+            <p>The app <strong>{{appName}}</strong> would like to use your {{siteTitle}} account to authenticate you.</p>
+            <p>If you allow this, <strong>{{appName}}</strong> will receive your name and email address.</p>
+            <br />
+
+            <form action="/oauth/authorize/decision" method="POST">
+                {{#if scopes.length}}
+                    <h3 class="title"><small>Additional data requested:</small></h3>
+                    <article class="message is-danger">
+                        <div class="message-body">{{error}}</div>
+                    </article>
+                    {{#each scopes}}
+                        <div class="field" style="text-align: left">
+                            <label class="label" for="scope-{{this.name}}">{{this.question}}</label>
+                            {{#if this.icon}}
+                                <div class="control has-icons-left">
+                                    <input class="input" type="{{this.type}}" id="scope-{{this.name}}" name="scope-{{this.name}}" {{{attr "value" this.value}}} required />
+                                    <span class="icon is-small is-left">
+                                        <i class="fas fa-{{this.icon}}"></i>
+                                    </span>
+                                </div>
+                            {{else}}
+                                <div class="control">
+                                    <input class="input" type="{{this.type}}" id="scope-{{this.name}}" name="scope-{{this.name}}" {{{attr "value" this.value}}} required />
+                                </div>
+                            {{/if}}
+                        </div>
+                    {{/each}}
+                    <br />
+                {{/if}}
+
+                <div class="field is-grouped is-grouped-centered">
+                    <input type="hidden" name="transaction_id" value="{{transactionID}}" />
+                    <p class="control">
+                        <input type="submit" class="button is-danger is-outlined" value="Cancel" name="cancel" formnovalidate />
+                    </p>
+                    <p class="control">
+                        <input type="submit" class="button is-success is-outlined" value="Allow" />
+                    </p>
+                </div>
+            </form>
+        </div>
+    </section>
+    {{/main}}
