+++ src/templates/resetpassword.hbs
@@ -1,0 +1,34 @@
+{{#> main }}
+    <article class="message is-danger">
+        <div class="message-body">{{error}}</div>
+    </article>
+    <article class="message is-success">
+        <div class="message-body">{{success}}</div>
+    </article>
+
+    <section class="columns">
+        <div class="column is-half is-offset-one-quarter">
+            <h1 class="title">Reset your password</h1>
+
+            <form action="/auth/forgot/{{resetCode}}" method="POST">
+                <div class="field">
+                    <label class="label" for="password1">New password</label>
+                    <div class="control">
+                        <input class="input" type="password" id="password1" name="password1" required />
+                    </div>
+                </div>
+                <div class="field">
+                    <label class="label" for="password2">Confirm password</label>
+                    <div class="control">
+                        <input class="input" type="password" id="password2" name="password2" required />
+                    </div>
+                </div>
+                <div class="field is-grouped is-grouped-centered">
+                    <p class="control">
+                        <input type="submit" class="button" value="Change password" />
+                    </p>
+                </div>
+            </form>
+        </div>
+    </section>
+    {{/main}}
