--- src/ui/changepassword.hbs
@@ -1,40 +1,0 @@
-{{#> main }}
-	<article class="message is-danger">
-		<div class="message-body">{{error}}</div>
-	</article>
-	<article class="message is-success">
-		<div class="message-body">{{success}}</div>
-	</article>
-
-	<section class="columns">
-		<div class="column is-half is-offset-one-quarter">
-			<h1 class="title">Change your password</h1>
-
-			<form action="/auth/changepassword" method="POST">
-				<div class="field">
-					<label class="label" for="oldpassword">Current password</label>
-					<div class="control">
-						<input class="input" type="password" id="oldpassword" name="oldpassword" required />
-					</div>
-				</div>
-				<div class="field">
-					<label class="label" for="password1">New password</label>
-					<div class="control">
-						<input class="input" type="password" id="password1" name="password1" required />
-					</div>
-				</div>
-				<div class="field">
-					<label class="label" for="password2">Confirm password</label>
-					<div class="control">
-						<input class="input" type="password" id="password2" name="password2" required />
-					</div>
-				</div>
-				<div class="field is-grouped is-grouped-centered">
-					<p class="control">
-						<input type="submit" class="button" value="Submit" />
-					</p>
-				</div>
-			</form>
-		</div>
-	</section>
-{{/main}}
