--- src/ui/forgotpassword.hbs
@@ -1,29 +1,0 @@
-{{#> main }}
-	<article class="message is-danger">
-		<div class="message-body">{{error}}</div>
-	</article>
-	<article class="message is-success">
-		<div class="message-body">{{success}}</div>
-	</article>
-
-	<section class="columns">
-		<div class="column is-half is-offset-one-quarter">
-			<h1 class="title">Reset your password</h1>
-			<p class="subtitle">Enter your email address and we'll send you a link to reset your password</p>
-
-			<form action="/auth/forgot" method="POST">
-				<div class="field">
-					<label class="label" for="email">Email</label>
-					<div class="control">
-						<input class="input" type="email" id="email" name="email" required />
-					</div>
-				</div>
-				<div class="field is-grouped is-grouped-centered">
-					<p class="control">
-						<input type="submit" class="button" value="Submit" />
-					</p>
-				</div>
-			</form>
-		</div>
-	</section>
-{{/main}}
