--- src/ui/index.hbs
@@ -1,33 +1,0 @@
-{{#> main }}
-	<section class="columns">
-		<div class="column is-half is-offset-one-quarter">
-			<h1 class="title">Hello, {{formatName user.name}}!</h1>
-			<p>You're logged in! Get the full details below:</p>
-			<p class="content">
-				<ul>
-					<li>Name: <strong>{{formatName user.name}}</strong></li>
-					<li>Email: <strong>{{user.email}}</strong></li>
-					<li>UUID: <strong>{{user.uuid}}</strong></li>
-					<li>Login method: <strong>{{loginMethod}}</strong></li>
-				</ul>
-			</p>
-
-			<br />
-			<div class="field is-grouped is-grouped-centered">
-				{{#if user.admin}}
-					<p class="control">
-						<a class="button is-medium" href="/admin">Admin Settings</a>
-					</p>
-				{{/if}}
-				{{#ifCond loginMethod "local"}}
-					<p class="control">
-						<a class="button is-medium" href="/login/changepassword">Change password</a>
-					</p>
-				{{/ifCond}}
-				<p class="control">
-					<a class="button is-medium" href="/logout">Log out</a>
-				</p>
-			</div>
-		</div>
-	</section>
-{{/main}}
