--- src/ui/tsconfig.json
@@ -1,12 +1,0 @@
-{
-	"compilerOptions": {
-		"module": "none",
-		"composite": true,
-		"target": "es5",
-		"lib": [
-			"dom"
-		],
-		"moduleResolution": "node",
-	},
-	"files": [ "login.ts", "admin.ts" ]
-}
